industry: separate code for distribution key from nodal production
This allows us to reuse the key for today's nodal energy demand from
industry.
